Refactoring. Implemented ISAAC OTP.

+++ libstrawberry/core/types/bool.h
@@ -37,7 +37,8 @@
 #include <stdint.h>
 
 
-typedef uint_fast32_t sb_bool_t;
+typedef uint_fast8_t sb_bool_t;
+
 
 #define sb_false							0
 #define sb_true								1
